:truck: Refactor #45: 폴더명 my_edison -> edison 수정
