+++ app/src/main/java/com/umc/edison/presentation/edison/BubbleInputState.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.presentation.my_edison
+package com.umc.edison.presentation.edison
 
 import com.umc.edison.presentation.label.LabelEditMode
 import com.umc.edison.presentation.model.BubbleModel
@@ -8,19 +8,18 @@ import com.umc.edison.ui.theme.Gray300
 data class BubbleInputState(
     val isLoading: Boolean = false,
     val bubble: BubbleModel,
-    val bubbles : List<BubbleModel>,
+    val bubbles: List<BubbleModel>,
     val labels: List<LabelModel>,
     val labelEditMode: LabelEditMode = LabelEditMode.NONE,
     val selectedLabel: LabelModel,
     val error: Throwable? = null,
-
-    ) {
+) {
     companion object {
         val DEFAULT = BubbleInputState(
             isLoading = false,
             bubble = BubbleModel(),
             bubbles = emptyList(),
-            labels =  emptyList(),
+            labels = emptyList(),
             labelEditMode = LabelEditMode.NONE,
             selectedLabel = LabelModel(
                 id = 0,
